+++ spec/unit/default_spec.rb
@@ -25,8 +25,8 @@ describe 'flywaydb::default' do
       expect(chef_run).to create_group('flyway')
     end
 
-    it 'creates directory' do
-      expect(chef_run).to create_directory("/opt/flyway-#{VERSION}")
+    it 'creates install directory' do
+      expect(chef_run).to create_directory('/opt/flywaydb')
     end
 
     it 'downloads flyway cli' do
@@ -39,6 +39,10 @@ describe 'flywaydb::default' do
 
     it 'untars flyway cli' do
       expect(chef_run).to_not run_execute('extract flyway')
+    end
+
+    it 'make flyway executable' do
+      expect(chef_run).to run_execute('chmod +x /opt/flywaydb/flyway/flyway')
     end
   end
 
@@ -62,8 +66,8 @@ describe 'flywaydb::default' do
       expect(chef_run).to create_group('Administrators')
     end
 
-    it 'creates directory' do
-      expect(chef_run).to create_directory("C:/flyway-#{VERSION}")
+    it 'creates install directory' do
+      expect(chef_run).to create_directory('C:/flywaydb')
     end
 
     it 'downloads flyway cli' do
@@ -76,6 +80,10 @@ describe 'flywaydb::default' do
 
     it 'unzips flyway cli' do
       expect(chef_run).to_not run_batch('unzip flyway (powershell 3 or higher required)')
+    end
+
+    it 'does not make flyway executable' do
+      expect(chef_run).to_not run_execute('chmod +x C:/flywaydb/flyway/flyway')
     end
   end
 end